refactor: rename AuthContext file to AuthProvider and update gitignore for agent artifacts
